requirements.txt
# Core dependencies
crewai>=0.28.8
langchain-groq>=0.1.0
groq>=0.4.0
python-dotenv>=1.0.0

# NLP dependencies
nltk>=3.8
textstat>=0.7.3
numpy>=1.24.0

# Utility dependencies
pydantic>=2.0.0
typing-extensions>=4.5.0

# Optional but recommended
requests>=2.28.0
json5>=0.9.0

# For development/testing
pytest>=7.0.0
black>=22.0.0


# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0


rich
streamlit
plotly
pandas

# Optional development tools
pytest==7.4.3 # for testing
black==23.11.0 # for code formatting